feat: gcs 오류 해결2

+++ src/main/java/com/pm/projectmanager/common/gcs/GoogleCloudStorageService.java
@@ -64,17 +64,32 @@ public class GoogleCloudStorageService {
         return gcsImageUrl;
     }
 
-    public boolean deleteImage(String fileName) throws IOException {
-        String keyFileName = "projectmanager_bucket_key.json"; // 인증키 파일
-        InputStream keyFile = getClass().getClassLoader().getResourceAsStream(keyFileName);
+    public boolean deleteImage(String fileName) {
+        try {
+            if (fileName == null || fileName.isEmpty()) {
+                return false;
+            }
 
-        Storage storage = StorageOptions.newBuilder()
-                .setCredentials(GoogleCredentials.fromStream(keyFile))
-                .build()
-                .getService();
-        String originFileName = extractFileNameFromGcsUrl(fileName);
-        return storage.delete(bucketName, originFileName);
+            String keyFileName = "projectmanager_bucket_key.json";
+            InputStream keyFile = getClass().getClassLoader().getResourceAsStream(keyFileName);
+
+            if (keyFile == null) {
+                return false;
+            }
+
+            Storage storage = StorageOptions.newBuilder()
+                    .setCredentials(GoogleCredentials.fromStream(keyFile))
+                    .build()
+                    .getService();
+
+            String originFileName = extractFileNameFromGcsUrl(fileName);
+            return storage.delete(bucketName, originFileName);
+
+        } catch (Exception e) {
+            return false;
+        }
     }
+
 
     public String extractFileNameFromGcsUrl(String url) {
         try {
